Move collection schema table constants and types out of component files
