+++ src/components/tables/Schema/Rows.tsx
@@ -1,8 +1,5 @@
-import type {
-    InferSchemaResponseProperty,
-    Schema,
-    SortDirection,
-} from 'src/types';
+import type { RowProps, RowsProps } from 'src/components/tables/Schema/types';
+import type { InferSchemaResponseProperty } from 'src/types';
 
 import { useMemo } from 'react';
 
@@ -12,21 +9,10 @@ import { orderBy } from 'lodash';
 
 import ChipListCell from 'src/components/tables/cells/ChipList';
 import { FieldList } from 'src/components/tables/cells/projections/FieldList';
+import { ROW_TYPE_STRING } from 'src/components/tables/Schema/shared';
 import { useEntityType } from 'src/context/EntityContext';
 import { useEntityWorkflow } from 'src/context/Workflow';
 import { basicSort_string } from 'src/utils/misc-utils';
-
-interface RowProps {
-    row: InferSchemaResponseProperty;
-}
-
-interface RowsProps {
-    data: Schema | null;
-    sortDirection: SortDirection;
-    columnToSort: string;
-}
-
-const rowTypeString = 'string';
 
 function Row({ row }: RowProps) {
     const entityType = useEntityType();
@@ -37,11 +23,11 @@ function Row({ row }: RowProps) {
     const formattedTypes = useMemo(() => {
         if (row.string_format) {
             const stringIndex = row.types.findIndex(
-                (rowType) => rowType === rowTypeString
+                (rowType) => rowType === ROW_TYPE_STRING
             );
             if (stringIndex > -1) {
                 row.types[stringIndex] =
-                    `${rowTypeString}: ${row.string_format}`;
+                    `${ROW_TYPE_STRING}: ${row.string_format}`;
             }
         }
